+++ src/main/webapp/WEB-INF/views/correcteurs.jsp
@@ -3,30 +3,35 @@
 <html>
 <head>
     <title>Liste Correcteurs</title>
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/css/style.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
 </head>
 <body>
-
-<h2>Liste des Correcteurs</h2>
- <a href="/correcteur/add">Ajouter un nouveau correcteur</a>
-<table border="1">
-<tr>
-    <th>ID</th>
-    <th>Nom</th>
-    <th>Action</th>
-</tr>
-
-<c:forEach var="c" items="${correcteurs}">
-<tr>
-    <td>${c.id}</td>
-    <td>${c.nom}</td>
-    <td>
-        <a href="/correcteur/edit/${c.id}">Modifier</a>
-        <a href="/correcteur/delete/${c.id}">Supprimer</a>
-    </td>
-</tr>
-</c:forEach>
-
-</table>
-<a href="/">← Retour à l’accueil</a>
+    <div class="app-layout">
+        <jsp:include page="menu.jsp" />
+        <main class="main-content">
+            <div class="container">
+                <h2>Liste des Correcteurs</h2>
+                <a href="${pageContext.request.contextPath}/correcteur/add" class="button">Ajouter un nouveau correcteur</a>
+                <table>
+                    <tr>
+                        <th>ID</th>
+                        <th>Nom</th>
+                        <th>Action</th>
+                    </tr>
+                    <c:forEach var="c" items="${correcteurs}">
+                        <tr>
+                            <td>${c.id}</td>
+                            <td>${c.nom}</td>
+                            <td>
+                                <a href="${pageContext.request.contextPath}/correcteur/edit/${c.id}">Modifier</a>
+                                <a href="${pageContext.request.contextPath}/correcteur/delete/${c.id}" onclick="return confirm('Confirmer la suppression ?')">Supprimer</a>
+                            </td>
+                        </tr>
+                    </c:forEach>
+                </table>
+            </div>
+        </main>
+    </div>
 </body>
 </html>
